Splitting informant,io_handler and webapps
